requirements.txt
joblib>= 0.16
scikit-learn>=0.23
pandas>= 1.0
numpy>=1.1
pyodbc==4.0.30
urllib3==1.25.9
sqlalchemy==1.3.18
pyyaml>=5.3
azureml-sdk>=1.18.0
